Вроде все работает, по тз все сделано, ща будем улудшать код

+++ Models/Weather.cs
@@ -6,14 +6,14 @@
         public DateOnly Date { get; set; } 
         public TimeSpan Time { get; set; }
         public double Temperature { get; set; }
-        public int Humidity { get; set; }
+        public double Humidity { get; set; }
         public double DewPoint { get; set; }
-        public int Pressure { get; set; }
+        public double Pressure { get; set; }
         public string WindDirection { get; set; }
         public double WindSpeed { get; set; }
-        public int CloudCover { get; set; }
-        public int LBCloudCover { get; set; }
-        public int HorizontalVisibility { get; set; }
+        public double CloudCover { get; set; }
+        public double LBCloudCover { get; set; }
+        public double HorizontalVisibility { get; set; }
         public string WeatherPhenomena { get; set; }
     }
 }
